Refactored unit test classes.
