+++ android/src/main/java/com/azure/reactnative/notificationhub/ReactNativeNotificationHubUtil.java
@@ -28,11 +28,9 @@ import org.json.JSONObject;
 import static com.azure.reactnative.notificationhub.ReactNativeConstants.*;
 
 public class ReactNativeNotificationHubUtil {
-    public static final String TAG = "NotificationHubUtil";
+    public static final String TAG = "ReactNativeNotificationHubUtil";
 
     private static ReactNativeNotificationHubUtil sharedNotificationHubUtilInstance = null;
-
-    private final ExecutorService mPool = Executors.newFixedThreadPool(1);
 
     private boolean mIsForeground;
 
@@ -161,194 +159,6 @@ public class ReactNativeNotificationHubUtil {
 
     public boolean getAppIsForeground() {
         return mIsForeground;
-    }
-
-    public void runInWorkerThread(Runnable runnable) {
-        mPool.execute(runnable);
-    }
-
-    public NotificationHub createNotificationHub(String hubName, String connectionString, ReactContext reactContext) {
-        NotificationHub hub = new NotificationHub(hubName, connectionString, reactContext);
-        return hub;
-    }
-
-    public JSONObject convertBundleToJSON(Bundle bundle) {
-        JSONObject json = new JSONObject();
-        Set<String> keys = bundle.keySet();
-        for (String key : keys) {
-            try {
-                json.put(key, bundle.get(key));
-            } catch (JSONException e) {
-            }
-        }
-
-        return json;
-    }
-
-    public Intent createBroadcastIntent(String action, JSONObject json) {
-        Intent intent = IntentFactory.createIntent(action);
-        intent.putExtra("event", DEVICE_NOTIF_EVENT);
-        intent.putExtra("data", json.toString());
-
-        return intent;
-    }
-
-    public Class getMainActivityClass(Context context) {
-        String packageName = context.getPackageName();
-        Intent launchIntent = context.getPackageManager().getLaunchIntentForPackage(packageName);
-        String className = launchIntent.getComponent().getClassName();
-        try {
-            return Class.forName(className);
-        } catch (ClassNotFoundException e) {
-            Log.e(TAG, ERROR_ACTIVITY_CLASS_NOT_FOUND, e);
-            return null;
-        }
-    }
-
-    public int getNotificationCompatPriority(String priorityString) {
-        int priority = NotificationCompat.PRIORITY_DEFAULT;
-        if (priorityString != null) {
-            switch (priorityString.toLowerCase()) {
-                case REMOTE_NOTIFICATION_PRIORITY_MAX:
-                    priority = NotificationCompat.PRIORITY_MAX;
-                    break;
-                case REMOTE_NOTIFICATION_PRIORITY_HIGH:
-                    priority = NotificationCompat.PRIORITY_HIGH;
-                    break;
-                case REMOTE_NOTIFICATION_PRIORITY_LOW:
-                    priority = NotificationCompat.PRIORITY_LOW;
-                    break;
-                case REMOTE_NOTIFICATION_PRIORITY_MIN:
-                    priority = NotificationCompat.PRIORITY_MIN;
-                    break;
-                case REMOTE_NOTIFICATION_PRIORITY_NORMAL:
-                    priority = NotificationCompat.PRIORITY_DEFAULT;
-                    break;
-            }
-        }
-
-        return priority;
-    }
-
-    public int getSmallIcon(Bundle bundle, Resources res, String packageName) {
-        int smallIconResId;
-        String smallIcon = bundle.getString(KEY_REMOTE_NOTIFICATION_SMALL_ICON);
-
-        if (smallIcon != null) {
-            smallIconResId = res.getIdentifier(smallIcon, RESOURCE_DEF_TYPE_MIPMAP, packageName);
-        } else {
-            smallIconResId = res.getIdentifier(RESOURCE_NAME_NOTIFICATION, RESOURCE_DEF_TYPE_MIPMAP, packageName);
-        }
-
-        if (smallIconResId == 0) {
-            smallIconResId = res.getIdentifier(RESOURCE_NAME_LAUNCHER, RESOURCE_DEF_TYPE_MIPMAP, packageName);
-
-            if (smallIconResId == 0) {
-                smallIconResId = android.R.drawable.ic_dialog_info;
-            }
-        }
-
-        return smallIconResId;
-    }
-
-    public int getLargeIcon(Bundle bundle, String largeIcon, Resources res, String packageName) {
-        int largeIconResId;
-
-        if (largeIcon != null) {
-            largeIconResId = res.getIdentifier(largeIcon, RESOURCE_DEF_TYPE_MIPMAP, packageName);
-        } else {
-            largeIconResId = res.getIdentifier(RESOURCE_NAME_LAUNCHER, RESOURCE_DEF_TYPE_MIPMAP, packageName);
-        }
-
-        return largeIconResId;
-    }
-
-    public Uri getSoundUri(Context context, Bundle bundle) {
-        Uri soundUri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
-        String soundName = bundle.getString(KEY_REMOTE_NOTIFICATION_SOUND_NAME);
-        if (soundName != null) {
-            if (!"default".equalsIgnoreCase(soundName)) {
-
-                // sound name can be full filename, or just the resource name.
-                // So the strings 'my_sound.mp3' AND 'my_sound' are accepted
-                // The reason is to make the iOS and android javascript interfaces compatible
-
-                int resId = context.getResources().getIdentifier(soundName, RESOURCE_DEF_TYPE_RAW, context.getPackageName());
-                if (resId == 0) {
-                    soundName = soundName.substring(0, soundName.lastIndexOf('.'));
-                    resId = context.getResources().getIdentifier(soundName, RESOURCE_DEF_TYPE_RAW, context.getPackageName());
-                }
-
-                soundUri = Uri.parse("android.resource://" + context.getPackageName() + "/" + resId);
-            }
-        }
-
-        return soundUri;
-    }
-
-    public Intent createNotificationIntent(Context context, Bundle bundle, Class intentClass) {
-        Intent intent = IntentFactory.createIntent(context, intentClass);
-        intent.addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
-        bundle.putBoolean(KEY_REMOTE_NOTIFICATION_FOREGROUND, true);
-        bundle.putBoolean(KEY_REMOTE_NOTIFICATION_USER_INTERACTION, false);
-        bundle.putBoolean(KEY_REMOTE_NOTIFICATION_COLDSTART, false);
-        intent.putExtra(KEY_INTENT_NOTIFICATION, bundle);
-
-        return intent;
-    }
-
-    public void processNotificationActions(Context context, Bundle bundle,
-                                           NotificationCompat.Builder notification,
-                                           int notificationID) {
-        JSONArray actionsArray = null;
-        try {
-            actionsArray = bundle.getString(KEY_REMOTE_NOTIFICATION_ACTIONS) != null ?
-                    new JSONArray(bundle.getString(KEY_REMOTE_NOTIFICATION_ACTIONS)) : null;
-        } catch (JSONException e) {
-            Log.e(TAG, ERROR_COVERT_ACTIONS, e);
-        }
-
-        if (actionsArray != null) {
-            // No icon for now. The icon value of 0 shows no icon.
-            int icon = 0;
-
-            // Add button for each actions.
-            for (int i = 0; i < actionsArray.length(); i++) {
-                String action;
-                try {
-                    action = actionsArray.getString(i);
-                } catch (JSONException e) {
-                    Log.e(TAG, ERROR_GET_ACTIONS_ARRAY, e);
-                    continue;
-                }
-
-                Intent actionIntent = IntentFactory.createIntent();
-                actionIntent.setAction(context.getPackageName() + "." + action);
-                // Add "action" for later identifying which button gets pressed.
-                bundle.putString(KEY_REMOTE_NOTIFICATION_ACTION, action);
-                actionIntent.putExtra(KEY_INTENT_NOTIFICATION, bundle);
-                PendingIntent pendingActionIntent = PendingIntent.getBroadcast(context, notificationID, actionIntent,
-                        PendingIntent.FLAG_UPDATE_CURRENT);
-                notification.addAction(icon, action, pendingActionIntent);
-            }
-        }
-    }
-
-    public NotificationCompat.Builder initNotificationCompatBuilder(Context context,
-                                                                    String notificationChannelID,
-                                                                    String title,
-                                                                    CharSequence ticker,
-                                                                    int visibility,
-                                                                    int priority,
-                                                                    boolean autoCancel) {
-        NotificationCompat.Builder notificationBuilder = new NotificationCompat.Builder(context, notificationChannelID)
-                .setContentTitle(title)
-                .setTicker(ticker)
-                .setVisibility(visibility)
-                .setPriority(priority)
-                .setAutoCancel(autoCancel);
-
-        return notificationBuilder;
     }
 
     private String getPref(Context context, String key) {
